Address PR review issues. Move TypeNameHelper to own file, rename to TypeExtensions and make internal. Tidy up ObjectToTypeName value converter.
