+++ sources/presentation/Xenko.Core.Presentation.Tests/TestValueConverters.cs
@@ -18,7 +18,6 @@ namespace Xenko.Core.Presentation.Tests
         {
             var converter = new ObjectToTypeName();
 
-
             Assert.Equal(converter.Convert(null, typeof(string), null, CultureInfo.CurrentCulture), ObjectToTypeName.NullObjectType);
         }
 
@@ -28,38 +27,6 @@ namespace Xenko.Core.Presentation.Tests
             var converter = new ObjectToTypeName();
 
             Assert.NotEqual(converter.Convert("hello", typeof(string), null, CultureInfo.CurrentCulture), ObjectToTypeName.NullObjectType);
-        }
-
-        [Fact]
-        public void TestTypeNameHelperReturnsCorrectlyFormattedName()
-        {
-            var tests = new List<(Type Type, string Name)>()
-            {
-                (typeof(int), "int"), //Simple type
-                (typeof(int?), "int?"), //Nullable simple type
-                (typeof(TimeSpan), "TimeSpan"), //Type
-                (typeof(TimeSpan?), "TimeSpan?"), //Nullable type
-                (typeof(int[]), "int[]"), //Array of simple type
-                (typeof(int?[]), "int?[]"), //Array of nullable simple type
-                (typeof(TimeSpan[]), "TimeSpan[]"), //Array of type
-                (typeof(TimeSpan?[]), "TimeSpan?[]"), //Array of nullable type
-                (typeof(TimeSpan[,,]), "TimeSpan[,,]"), //Multi-dimesional array of type
-                (typeof(Dictionary<string, FactAttribute>), "Dictionary<string, FactAttribute>"), //Generic type
-                (typeof((string, FactAttribute)), "(string, FactAttribute)"), //Tuple types
-                (typeof((string, FactAttribute)?), "(string, FactAttribute)?"), //Nullable tuple types
-                (typeof(Dictionary<string, GenericStruct<int?>>), "Dictionary<string, GenericStruct<int?>>"), //Nested generic type
-                (typeof((GenericStruct<int?>?, double)), "(GenericStruct<int?>?, double)"), //Crazy type
-            };
-
-            foreach (var item in tests)
-            {
-                Assert.Equal(item.Type.ToSimpleCSharpName(), item.Name);
-            }            
-        }
-
-        private struct GenericStruct<T>
-        {
-            public T Field;
         }
     }
 }
